refactor(engine): Make get_activites a BaseService classmethod
